Struct getargs::Options

pub struct Options<A: Argument, I: Iterator<Item = A>> { /* private fields */ }

An argument parser.

Options accepts an iterator of Arguments and transforms them into Opts or Args, their values, and positional arguments.

First, create an iterator over your arguments and call Options::new to create a new Options.

You call Options::next_opt or Options::next_arg to attempt to parse the next argument as an Opt or Arg respectively. These methods can return an Error if the last option was provided with a value but did not consume it. They can also return Ok(None) if the next argument is not an option, or if the end of the arguments list has been reached.

Any returned error can be safely ignored and parsing will continue as normal, however, this is not generally recommended.

After a call to next_* returns an Opt (or an Arg::Short/Arg::Long), you may call either Options::value or Options::value_opt to retrieve the value of the option. In the absence of any call to a value-consuming method, getargs assumes the option must not have any value, and will raise that as an error in the next call to next_*.

Once next_* returns Ok(None), you may call Options::opts_ended to check if -- was used. It will never panic.

You may call Options::next_positional, Options::positionals, or Options::into_positionals whenever an option is not currently being parsed. (You can call Options::value_opt to finish the parsing of the current option if the last call to next_* returned an option, but be aware that the returned value will not be included in the arguments due to being a part of the option.)

See the crate-level documentation for a detailed explanation with step-by-step code examples.

Examples

Here is an idiomatic explainer:

let args = ["--help"];
let mut opts = Options::new(args.into_iter());

while let Some(opt) = opts.next_opt().expect("argument parsing error") {
    // We now have a call to `Options::value` / `Options::value_str`
    // to declare whether or not `opt` takes a value. In the absence
    // of any such call, `getargs` assumes there cannot be a value.

    match opt {
        Opt::Short('h') | Opt::Long("help") => { /* print help... */ }
        Opt::Short('v') => { /* print version, or increment verbosity, or... */ }
        Opt::Short('e') => {
            // The error for `Options::value` includes the option
            let value = opts.value().unwrap();

            // do something with value...
        }

        // `Opt` implements `Display` (if its argument does)
        _ => panic!("unknown option: {}", opt)
    }
}

// `Options::opts_ended` could be used here to detect `--`

// could easily be `while let Some(arg) = opts.next_positional()`
for arg in opts.positionals() {
    // do something with each positional argument...

    // If you are concerned about flags being passed here, see the
    // `anywhere` example, which allows flags anywhere
}

Here are a couple more examples in unit-test format:

let args = ["-a", "--bee", "foo"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('a'))));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("bee"))));
assert_eq!(opts.next_opt(), Ok(None));

// Don't forget the positional arguments!
assert_eq!(opts.next_positional(), Some("foo"));
assert_eq!(opts.next_positional(), None);

Retrieving values from options:

let args = ["--cupcake=value", "-b", "value2", "-v", "foo"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("cupcake"))));
assert_eq!(opts.value(), Ok("value"));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('b'))));
assert_eq!(opts.value(), Ok("value2"));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('v'))));
assert_eq!(opts.next_opt(), Ok(None));
assert_eq!(opts.next_positional(), Some("foo"));
assert_eq!(opts.next_positional(), None);

Parsing options anywhere:

let args = ["x", "-r", "hi", "y", "-ohi", "z", "--", "--not-a-flag"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("x"))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Short('r'))));
assert_eq!(opts.value(), Ok("hi"));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("y"))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Short('o'))));
assert_eq!(opts.value_opt(), Some("hi"));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("z"))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("--not-a-flag"))));
assert_eq!(opts.next_arg(), Ok(None));

Implementations

impl<'arg, A: Argument + 'arg, I: Iterator<Item = A>> Options<A, I>

pub fn new(iter: I) -> Options<A, I>

Creates a new Options given an iterator over arguments of type A.

The argument parser only lives as long as the iterator, but returns arguments with the same lifetime as whatever the iterator yields.

pub fn next_opt(&mut self) -> Result<A, Option<Opt<A>>>

Retrieves the next option.

Returns Ok(None) if there are no more options, or Err(..) if a parse error occurs.

This method also does not retrieve any value that goes with an option. If the option requires an value, such as in --option=value, then you should call Options::value after receiving the option from this method. Alternatively, you can call Options::value_opt instead for optional values.

Once this method returns Ok(None), then you should make sure to use Options::next_positional or Options::positionals to get the values of the following positional arguments. See the Options documentation for more details.

If your application accepts positional arguments in between flags, you can use Options::next_arg instead of next_opt.

pub fn next_arg(&mut self) -> Result<A, Option<Arg<A>>>

Retrieves the next argument. An argument is represented by Arg and defined as either an option (Opt) or a positional argument.

This method is implemented entirely using other public methods of Options.

Panics

Panics if Options::next_opt does.

Example

let args = ["--option", "VALUE", "arg", "arg2", "-vvv", "--option", "VALUE", "--", "arg3", "arg4", "--not-an-option"];
let mut opts = Options::new(args.into_iter());

// Regular options and values work...
assert_eq!(opts.next_arg(), Ok(Some(Arg::Long("option"))));
assert_eq!(opts.value(), Ok("VALUE"));

// Positional arguments afterwards also work...
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("arg"))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("arg2"))));

// Options after positional arguments also work
assert_eq!(opts.next_arg(), Ok(Some(Arg::Short('v'))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Short('v'))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Short('v'))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Long("option"))));
assert_eq!(opts.value(), Ok("VALUE"));

// `--` is recognized and handled correctly
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("arg3"))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("arg4"))));
assert_eq!(opts.next_arg(), Ok(Some(Arg::Positional("--not-an-option"))));
assert_eq!(opts.next_arg(), Ok(None));

pub fn value(&mut self) -> Result<A, A>

Retrieves the value passed to the option last returned by Options::next_opt or Options::next_arg.

This function returns an error if there is no value to return because the end of the argument list has been reached.

Panics

This method panics if Options::next_opt or Options::next_arg have not yet been called, if value is called twice for the same option, or if the last call to next_* did not return an option.

Example

let args = ["-aay", "--bee=foo", "-c", "see", "bar"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('a'))));
assert_eq!(opts.value(), Ok("ay"));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("bee"))));
assert_eq!(opts.value(), Ok("foo"));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('c'))));
assert_eq!(opts.value(), Ok("see"));

pub fn value_opt(&mut self) -> Option<A>

Retrieves an optional value for the option last returned by Options::next_opt or Options::next_arg. Only explicit values are accepted (--flag=VALUE, -fVALUE). Implicit values (--flag VALUE, -f VALUE) will not be consumed, and they will return None.

The reasoning for this is that if implicit values were consumed by this method, a subsequent flag could be mistaken as the value. --opt= also needs to be distinct from --opt - the former needs to have an empty value, whereas the latter needs to have no value.

Short options do not support empty values.

Panics

This method panics if Options::next_opt or Options::next_arg have not yet been called, if value_opt is called twice for the same option, or if the last call to next_* did not return an option.

Example

let args = ["--opt=value", "--other-flag", "--opt", "--opt=other", "-o", "-ovalue"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("opt"))));
assert_eq!(opts.value_opt(), Some("value"));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("other-flag"))));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("opt"))));
assert_eq!(opts.value_opt(), None); // does not return "--opt=other"
assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("opt"))));
assert_eq!(opts.value_opt(), Some("other"));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('o'))));
assert_eq!(opts.value_opt(), None); // short options can't have empty values
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('o'))));
assert_eq!(opts.value_opt(), Some("value"));
assert_eq!(opts.next_opt(), Ok(None));

pub fn next_positional(&mut self) -> Option<A>

Retrieves the next positional argument. This method must be called after all the last option has been fully parsed. Usually, this is when Options::next_opt returns Ok(None), but it can also be after calling Options::value_opt.

After this method is called, you can parse further options with Options::next_opt or continue getting positional arguments (which will not treat options specially).

This method preserves the state of Options::opts_ended.

Panics

This method panics if the option parsing is not yet complete; that is, it can panic if Options::next_opt has not yet returned None at least once. If Options::next_opt returns None, then Options::next_positional will always be safe to call.

Examples

Basic usage:

let args = ["-a", "foo", "bar"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('a'))));
assert_eq!(opts.next_opt(), Ok(None));
assert_eq!(opts.next_positional(), Some("foo"));
assert_eq!(opts.next_positional(), Some("bar"));
assert_eq!(opts.next_positional(), None);

pub fn positionals(&mut self) -> Positionals<'_, A, I>

Returns an iterator over the positional arguments of this Options. The returned iterator will forward Iterator::next calls to Options::next_positional.

This method does not panic, but Iterator::next may panic once it is called if option parsing has not finished (Options::next_opt has not returned Ok(None)). The exact requirements are the same as Options::next_positional.

Example

let args = ["--flag", "one", "two"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_opt(), Ok(Some(Opt::Long("flag"))));
assert_eq!(opts.next_opt(), Ok(None));

let mut args = opts.positionals();

assert_eq!(args.next(), Some("one"));
assert_eq!(args.next(), Some("two"));
assert_eq!(args.next(), None);

pub fn into_positionals(self) -> IntoPositionals<A, I>

Consumes this Options, returning an iterator over the rest of the arguments. The returned iterator wraps the one originally passed to Options::new.

Panics

Panics if an option is currently being parsed.

pub fn opts_ended(&self) -> bool

Returns true if the last call to Options::next_opt encountered a --. In that case, it will have returned None, but without this method you wouldn’t be able to tell whether that was due to the -- or simply the next argument being positional (or not existing).

This flag is not reset by Options::next_positional.

Example

let args = ["-a", "-b", "foo", "-c", "-d", "--", "bar", "-e", "-f"];
let mut opts = Options::new(args.into_iter());

assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('a'))));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('b'))));
assert_eq!(opts.next_opt(), Ok(None));
assert_eq!(opts.next_positional(), Some("foo"));

// `Options::next_opt` returned `None` because of `foo` (non-option)
assert!(!opts.opts_ended());
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('c'))));
assert_eq!(opts.next_opt(), Ok(Some(Opt::Short('d'))));
assert_eq!(opts.next_opt(), Ok(None));

// `Options::next_opt` returned `None` because of `--`
assert!(opts.opts_ended());

// Even after an argument is consumed, `opts_ended` remembers
assert_eq!(opts.next_positional(), Some("bar"));
assert!(opts.opts_ended());
assert_eq!(opts.next_positional(), Some("-e"));
assert_eq!(opts.next_positional(), Some("-f"));
assert_eq!(opts.next_positional(), None);
assert!(opts.opts_ended());

pub fn reset_opts_ended(&mut self)

Resets Options::opts_ended, causing it to start returning false. This is useful if you have multiple “levels” of options and are using Options::next_arg rather than Options::next_opt directly.

Panics

This function does not panic. If executed in an invalid state, it is simply a no-op.

pub fn is_empty(&self) -> bool

Returns true if this Options has reached the end of its iterator, and all positional arguments have also been consumed. This method will always return true when Options::next_positional does.

pub fn restart(&mut self)

“Restarts” options parsing if the iterator has been exhausted (Options::next_positional returned None). This only results in any noticeable effect if the iterator is a repeating iterator; otherwise, the iterator will never produce more arguments anyway.

Most iterators, including iterators over arrays, never repeat their elements. This method isn’t very useful unless you know the exact behavior of the iterator you’re using.

Panics

Panics if iteration is not over, as defined by Options::is_empty.